# docs/engine-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Engine de Canonização — API",
    "description": "Canonização de itens de cupom fiscal brasileiro: descrição suja → produto canônico + marca + confiança por item. Pipeline determinístico (EAN, dicionário de abreviações, base de conhecimento compartilhada, fuzzy, regex) com IA apenas para a cauda longa — cada item custa IA no máximo uma vez, globalmente. Autenticação atual: Bearer JWT (playground); versão B2B usará API key por cliente.",
    "version": "0.1.0",
    "contact": {
      "name": "CookMe"
    }
  },
  "servers": [
    {
      "url": "https://api.cookme.com.br",
      "description": "Produção"
    },
    {
      "url": "http://localhost:3000",
      "description": "Dev"
    }
  ],
  "paths": {
    "/api/engine/canonizar": {
      "post": {
        "description": "Recebe descrições sujas de itens de cupom fiscal brasileiro e devolve, por item: nome canônico do produto, marca, flag de alimento, confiança (0–1) e o estágio do pipeline que resolveu (EAN → dicionário → base de conhecimento → fuzzy → regex → IA → fallback). Itens que a cadeia determinística não resolve com confiança passam por IA e o resultado é aprendido na base compartilhada — o mesmo item nunca custa IA duas vezes.",
        "operationId": "EngineController_canonizar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CanonizarRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonizarResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Canoniza itens de cupom fiscal",
        "tags": [
          "Engine — Canonização"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ItemEntradaDto": {
        "type": "object",
        "properties": {
          "descricao": {
            "type": "string",
            "description": "Linha do item como veio do cupom fiscal (OCR ou QR)",
            "example": "CR LEITE ITALAC 200GR"
          },
          "ean": {
            "type": "string",
            "description": "EAN/GTIN quando disponível — resolve com confiança máxima",
            "example": "7898080640611"
          },
          "preco": {
            "type": "number",
            "description": "Preço unitário em reais",
            "example": 3.49
          },
          "quantidade": {
            "type": "number",
            "description": "Quantidade comprada",
            "example": 2
          },
          "unidade": {
            "type": "string",
            "description": "Unidade (kg, un, l…)",
            "example": "un"
          }
        },
        "required": [
          "descricao"
        ]
      },
      "CanonizarRequestDto": {
        "type": "object",
        "properties": {
          "itens": {
            "description": "Itens a canonizar — strings cruas ou objetos. Máximo 100 por chamada.",
            "example": [
              "CR LEITE ITALAC 200GR",
              {
                "descricao": "FILE PEITO FGO CONG SEARA KG",
                "preco": 18.9
              },
              "SAB EM PO OMO 1KG"
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemEntradaDto"
            }
          }
        },
        "required": [
          "itens"
        ]
      },
      "ItemCanonizadoDto": {
        "type": "object",
        "properties": {
          "descricao_original": {
            "type": "string",
            "example": "CR LEITE ITALAC 200GR"
          },
          "produto_canonico": {
            "type": "string",
            "description": "Nome canônico do produto: minúsculas, sem marca, sem peso/embalagem",
            "example": "creme de leite"
          },
          "marca": {
            "type": "object",
            "description": "Marca detectada na descrição, ou null",
            "example": "Italac",
            "nullable": true
          },
          "eh_alimento": {
            "type": "object",
            "description": "false = certeza de não-alimento (guard determinístico); null = indeterminado",
            "example": null,
            "nullable": true
          },
          "ean": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "preco": {
            "type": "object",
            "example": 3.49,
            "nullable": true
          },
          "quantidade": {
            "type": "object",
            "example": 2,
            "nullable": true
          },
          "confianca": {
            "type": "number",
            "description": "Confiança honesta por estágio de resolução: ean 0.99 · dicionario 0.95 · kb 0.92 · regex 0.88 · fuzzy 0.75 · ia 0.7 · normalizer 0.55 · fallback 0.3",
            "example": 0.95,
            "minimum": 0,
            "maximum": 1
          },
          "estagio": {
            "type": "string",
            "description": "Estágio do pipeline que resolveu o item",
            "enum": [
              "ean",
              "dicionario",
              "kb",
              "fuzzy",
              "regex",
              "normalizer",
              "ia",
              "fallback"
            ],
            "example": "dicionario"
          }
        },
        "required": [
          "descricao_original",
          "produto_canonico",
          "marca",
          "eh_alimento",
          "ean",
          "preco",
          "quantidade",
          "confianca",
          "estagio"
        ]
      },
      "CanonizarResponseDto": {
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemCanonizadoDto"
            }
          },
          "total": {
            "type": "number",
            "example": 3
          },
          "latencia_ms": {
            "type": "number",
            "description": "Tempo de processamento do lote em ms",
            "example": 17
          },
          "confianca_media": {
            "type": "number",
            "description": "Média das confianças do lote",
            "example": 0.95
          }
        },
        "required": [
          "itens",
          "total",
          "latencia_ms",
          "confianca_media"
        ]
      }
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}